+++ example/lib/screens/component/google_pay/google_pay_advanced_component_screen.dart
@@ -1,7 +1,5 @@
 // ignore_for_file: unused_local_variable
 
-import 'dart:convert';
-
 import 'package:adyen_checkout/adyen_checkout.dart';
 import 'package:adyen_checkout_example/config.dart';
 import 'package:adyen_checkout_example/repositories/adyen_google_pay_component_repository.dart';
@@ -33,12 +31,12 @@ class GooglePayAdvancedComponentScreen extends StatelessWidget {
   }
 
   Widget _buildAdyenGooglePayAdvancedComponent() {
-    return FutureBuilder<String>(
+    return FutureBuilder<Map<String, dynamic>>(
       future: repository.fetchPaymentMethods(),
-      builder: (BuildContext context, AsyncSnapshot<String> snapshot) {
+      builder:
+          (BuildContext context, AsyncSnapshot<Map<String, dynamic>> snapshot) {
         if (snapshot.hasData) {
-          final AdvancedCheckout advancedCheckout =
-              AdvancedCheckout(
+          final AdvancedCheckout advancedCheckout = AdvancedCheckout(
             onSubmit: repository.onSubmit,
             onAdditionalDetails: repository.onAdditionalDetails,
           );
@@ -93,13 +91,13 @@ class GooglePayAdvancedComponentScreen extends StatelessWidget {
     );
   }
 
-  Map<String, dynamic> _extractPaymentMethod(String paymentMethods) {
+  Map<String, dynamic> _extractPaymentMethod(
+      Map<String, dynamic> paymentMethods) {
     if (paymentMethods.isEmpty) {
       return <String, String>{};
     }
 
-    Map<String, dynamic> jsonPaymentMethods = jsonDecode(paymentMethods);
-    return jsonPaymentMethods["paymentMethods"].firstWhere(
+    return paymentMethods["paymentMethods"].firstWhere(
       (paymentMethod) => paymentMethod["type"] == "googlepay",
       orElse: () => throw Exception("Google pay payment method not provided"),
     );